File manager

+++ documents/trashbin/trashbin.controllers.js
@@ -1,12 +1,19 @@
 import asyncHandler from "express-async-handler";
 import { PrismaClient } from "@prisma/client";
-import {
- deleteFile
-} from "../../utils/services/awsS3bucket.js";
+import { deleteFile } from "../../utils/services/awsS3bucket.js";
+import deleteFolderRecursive from "../../utils/helpers/deleteChildFolders.js";
 const prisma = new PrismaClient();
 
-// TODO: empty trashbin: this will delete all folders and files in the trashbin
-// TODO: delete some files and folders in the trashbin
+// Get all trashbins in a workspace
+const getAllTrashbins = asyncHandler(async (req, res) => {
+  try {
+    const trashbins = await prisma.trashBin.findMany({});
+
+    res.status(200).json(trashbins);
+  } catch (error) {
+    res.status(500).json({ error: error });
+  }
+});
 
 //  move folders and files to trash bin
 const moveFoldersAndFilesToTrash = asyncHandler(async (req, res) => {
@@ -189,6 +196,88 @@ const deleteSelectedFoldersAndFiles = asyncHandler(async (req, res) => {
   try {
     const { folderIds, fileIds } = req.body;
 
+    // Check if folderIds are provided
+    if (!folderIds || folderIds.length === 0) {
+      const existingFiles = await prisma.file.findMany({
+        where: {
+          id: {
+            in: fileIds,
+          },
+        },
+      });
+
+      // If any file in the trash bin doesn't exist, return an error
+      if (existingFiles.length !== fileIds.length) {
+        return res.status(404).json({
+          error: "files already deleted from database",
+        });
+      }
+
+      // If no folderIds are provided, proceed with deleting only files
+      for (const fileId of fileIds) {
+        const file = await prisma.file.findUnique({
+          where: { id: fileId },
+          select: { fileIdentifier: true },
+        });
+
+        if (file) {
+          await deleteFile(file.fileIdentifier);
+        }
+      }
+
+      // Delete mappings for selected files
+      await prisma.fileFolderMapping.deleteMany({
+        where: {
+          fileId: {
+            in: fileIds,
+          },
+        },
+      });
+      // Delete files from database
+      await prisma.file.deleteMany({
+        where: {
+          id: {
+            in: fileIds,
+          },
+        },
+      });
+      return res.status(200).json({
+        msg: "Selected files deleted successfully",
+      });
+    }
+
+    // Check if all folders and files in the trash bin exist
+    const existingFolders = await prisma.folder.findMany({
+      where: {
+        id: {
+          in: folderIds,
+        },
+      },
+    });
+
+    const existingFiles = await prisma.file.findMany({
+      where: {
+        id: {
+          in: fileIds,
+        },
+      },
+    });
+
+    // If any folder or file in the trash bin doesn't exist, return an error
+    if (
+      existingFolders.length !== folderIds.length ||
+      existingFiles.length !== fileIds.length
+    ) {
+      return res
+        .status(404)
+        .json({ error: "Some folders or files in the trash bin do not exist" });
+    }
+
+    // Delete selected folders
+    for (const folderId of folderIds) {
+      await deleteFolderRecursive(folderId);
+    }
+
     // Delete selected folders
     await prisma.folder.deleteMany({
       where: {
@@ -198,6 +287,19 @@ const deleteSelectedFoldersAndFiles = asyncHandler(async (req, res) => {
       },
     });
 
+    // Delete files from AWS S3
+    for (const fileId of fileIds) {
+      // Get the file details to retrieve the file name
+      const file = await prisma.file.findUnique({
+        where: { id: fileId },
+        select: { fileIdentifier: true },
+      });
+
+      if (file) {
+        // Delete the file from AWS S3
+        await deleteFile(file.fileIdentifier);
+      }
+    }
     // Delete selected files using FileFolderMapping
     await prisma.fileFolderMapping.deleteMany({
       where: {
@@ -207,7 +309,116 @@ const deleteSelectedFoldersAndFiles = asyncHandler(async (req, res) => {
       },
     });
 
-      // Delete files from AWS S3
+    res.status(200).json({
+      msg: "Selected folders and files deleted successfully",
+    });
+  } catch (error) {
+    console.error(error);
+    res.status(500).json({ error: error.message });
+  }
+});
+
+const emptyTrashBin = asyncHandler(async (req, res) => {
+  try {
+    const employeeId = req.employeeId;
+
+    // Find all trash bin entries
+    const trashBinEntries = await prisma.trashBin.findMany({
+      where: {
+        employeeId,
+      },
+    });
+
+    // If the trash bin is already empty, return an error message
+    if (trashBinEntries.length === 0) {
+      return res.status(404).json({ error: "The trash bin is already empty" });
+    }
+
+    // Separate folder IDs  from the trash bin entries
+    const folderIds = trashBinEntries
+      .filter((entry) => entry.folderId)
+      .map((entry) => entry.folderId);
+
+    // Separate file IDs from the trash bin entries
+    const fileIds = trashBinEntries
+      .filter((entry) => entry.fileId)
+      .map((entry) => entry.fileId);
+
+    // Check if all files in the trash bin exist
+    const allFiles = await prisma.file.findMany({
+      where: {
+        id: {
+          in: fileIds,
+        },
+      },
+    });
+
+    // If any file in the trash bin doesn't exist, return an error
+    if (allFiles.length !== fileIds.length) {
+      return res.status(404).json({
+        error: "Some files in the trash bin do not exist",
+      });
+    }
+
+    // Check if folderIds are provided
+    if (!folderIds || folderIds.length === 0) {
+      // If no folderIds are provided, proceed with deleting only files
+      for (const fileId of fileIds) {
+        const file = await prisma.file.findUnique({
+          where: { id: fileId },
+          select: { fileIdentifier: true },
+        });
+
+        if (file) {
+          await deleteFile(file.fileIdentifier);
+        }
+      }
+
+      // Delete mappings for selected files
+      await prisma.fileFolderMapping.deleteMany({
+        where: {
+          fileId: {
+            in: fileIds,
+          },
+        },
+      });
+
+      // Delete files from database
+      await prisma.file.deleteMany({
+        where: {
+          id: {
+            in: fileIds,
+          },
+        },
+      });
+
+      // Delete all trash bin entries
+      await prisma.trashBin.deleteMany({
+        where: {
+          employeeId,
+        },
+      });
+
+      return res.status(200).json({
+        msg: "Selected files deleted successfully",
+      });
+    }
+
+    // Delete selected folders and associated files
+    for (const folderId of folderIds) {
+      await deleteFolderRecursive(folderId);
+    }
+
+    // Delete selected folders
+    await prisma.folder.deleteMany({
+      where: {
+        id: {
+          in: folderIds,
+        },
+      },
+    });
+
+    // Delete files from AWS S3
     for (const fileId of fileIds) {
       // Get the file details to retrieve the file name
       const file = await prisma.file.findUnique({
@@ -221,118 +432,19 @@ const deleteSelectedFoldersAndFiles = asyncHandler(async (req, res) => {
       }
     }
 
-    res.status(200).json({
-      msg: "Selected folders and files deleted successfully",
-    });
-  } catch (error) {
-    console.error(error);
-    res.status(500).json({ error: error.message });
-  }
-});
-
-const emptyTrashBin = asyncHandler(async (req, res) => {
-  try {
-    const { workspaceId } = req.params;
-    const employeeId = req.employeeId;
-    // Find all trash bin entries
-    const trashBinEntries = await prisma.trashBin.findMany({
+    // Delete mappings for selected files
+    await prisma.fileFolderMapping.deleteMany({
       where: {
-        workspaceId,
-        employeeId,
+        fileId: {
+          in: fileIds,
+        },
       },
     });
-
-    // Separate folder IDs and file IDs from the trash bin entries
-    const folderIds = trashBinEntries
-      .filter((entry) => entry.folderId)
-      .map((entry) => entry.folderId);
-    const fileIds = trashBinEntries
-      .filter((entry) => entry.fileId)
-      .map((entry) => entry.fileId);
-
-    // // Check if all folders and files in the trash bin exist
-    // const existingFolders = await prisma.folder.findMany({
-    //   where: {
-    //     id: {
-    //       in: folderIds,
-    //     },
-    //   },
-    // });
-
-    // const existingFiles = await prisma.file.findMany({
-    //   where: {
-    //     id: {
-    //       in: fileIds,
-    //     },
-    //   },
-    // });
-
-    // // If any folder or file in the trash bin doesn't exist, return an error
-    // if (
-    //   existingFolders.length !== folderIds.length ||
-    //   existingFiles.length !== fileIds.length
-    // ) {
-    //   return res
-    //     .status(404)
-    //     .json({ error: "Some folders or files in the trash bin do not exist" });
-    // }
 
     // Delete all trash bin entries
     await prisma.trashBin.deleteMany({
       where: {
-        workspaceId,
         employeeId,
-      },
-    });
-
-    // Delete associated file-folder mappings
-    await prisma.fileFolderMapping.deleteMany({
-      where: {
-        OR: [
-          {
-            fileId: {
-              in: fileIds,
-            },
-          },
-          {
-            folderId: {
-              in: folderIds,
-            },
-          },
-        ],
-      },
-    });
-
-
-// Delete the files from AWS S3
-
-    // Delete files from AWS S3
-    for (const fileId of fileIds) {
-      // Get the file details to retrieve the file name
-      const file = await prisma.file.findUnique({
-        where: { id: fileId },
-        select: { fileIdentifier: true },
-      });
-
-      if (file) {
-        await deleteFile(file.fileIdentifier); 
-      }
-    }
-
-    // Delete all associated files and folders
-    await prisma.folder.deleteMany({
-      where: {
-        id: {
-          in: folderIds,
-        },
-      },
-    });
-
-    await prisma.file.deleteMany({
-      where: {
-        id: {
-          in: fileIds,
-        },
       },
     });
 
@@ -350,4 +462,5 @@ export {
   restoreFoldersAndFilesFromTrash,
   deleteSelectedFoldersAndFiles,
   emptyTrashBin,
+  getAllTrashbins,
 };
